### 2023-09-02 #5801 @editorialbot
[PRE REVIEW]: PerMetrics: A Framework of Performance Metrics for Machine Learning Models
**Submitting author:** @thieu1995 (<a href="http://orcid.org/0000-0001-9994-8747">Nguyen Van Thieu</a>)
**Repository:** https://github.com/thieu1995/permetrics
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.4.2
