+++ paper.bib
@@ -137,7 +137,8 @@
   booktitle={2021 IEEE International Conference on Services Computing (SCC)},
   pages={450--455},
   year={2021},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/scc53864.2021.00065},
 }
 
 @article{harris2020array,
@@ -148,7 +149,9 @@
   number={7825},
   pages={357--362},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-020-2649-2},
+  url={https://doi.org/10.1038/s41586-020-2649-2},
 }
 
 @article{virtanen2020scipy,
@@ -159,7 +162,8 @@
   number={3},
   pages={261--272},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-019-0686-2},
 }
 
 @article{scikit_learn,
@@ -178,15 +182,15 @@
    author = {Ben Hamner},
    title = {Metrics},
    year = 2015,
-   url = "https://github.com/benhamner/Metrics"
+   url = {https://github.com/benhamner/Metrics},
 }
 
 @software{torchmetrics,
+	title = {{TorchMetrics - Measuring Reproducibility in PyTorch}},
 	author = {{Nicki Skafte Detlefsen} and {Jiri Borovec} and {Justus Schock} and {Ananya Harsh} and {Teddy Koker} and {Luca Di Liello} and {Daniel Stancl} and {Changsheng Quan} and {Maxim Grechkin} and {William Falcon}},
 	doi = {10.21105/joss.04101},
 	license = {Apache-2.0},
 	month = feb,
-	title = {{TorchMetrics - Measuring Reproducibility in PyTorch}},
 	url = {https://github.com/Lightning-AI/torchmetrics},
 	year = {2022}
 }
@@ -195,8 +199,11 @@
 	title={Tensorflow: Large-scale machine learning on heterogeneous distributed systems},
 	author={Abadi, Mart{\'\i}n and Agarwal, Ashish and Barham, Paul and Brevdo, Eugene and Chen, Zhifeng and Citro, Craig and Corrado, Greg S and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and others},
 	journal={arXiv preprint arXiv:1603.04467},
+	url={https://doi.org/10.48550/arXiv.1603.04467},
+	doi={10.48550/arXiv.1603.04467},
 	year={2016}
 }
+
 
 @misc{chollet2015keras,
 	title={Keras},
@@ -211,14 +218,16 @@
   journal={Advances in neural information processing systems},
   volume={31},
   year={2018},
-  url={https://catboost.ai}
+  url={https://catboost.ai},
+  doi={10.48550/arXiv.1706.09516},
 }
 
 @article{chen2015mxnet,
   title={Mxnet: A flexible and efficient machine learning library for heterogeneous distributed systems},
   author={Chen, Tianqi and Li, Mu and Li, Yutian and Lin, Min and Wang, Naiyan and Wang, Minjie and Xiao, Tianjun and Xu, Bing and Zhang, Chiyuan and Zhang, Zheng},
   journal={arXiv preprint arXiv:1512.01274},
-  year={2015}
+  year={2015},
+  doi={10.48550/arXiv.1512.01274},
 }
 
 @article{desgraupes2013clustering,
