### 2023-09-10 #5801 @thieu1995
Hi @arfon ,

Could you please provide me with the reason you consider my submission as out-of-scope? I have noticed you providing reasons for other submissions, but not for mine. Additionally, you mentioned adding 'query-scope' to my submission. Can you clarify this further?
