### 2023-09-11 #5801 @danielskatz
👋 @thieu1995 - @arfon did not say this is out of scope, but that the editors will consider if it is or not. The question is if it is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS
