### 2023-09-11 #5801 @thieu1995
Hi @danielskatz ,

Thank you for clarifying my question. I will await the editors' decision. However, I would like to mention that my library could be categorized as supporting the functioning of research instruments or the execution of research experiments, as well as offering a mathematical library. It is used by numerous researchers and users from different domains, with the current number of downloads being around 36k (https://www.pepy.tech/projects/permetrics).
Of course, the final decisions rest with the editors. Thank you.
