### 2023-12-12 #6143 @editorialbot
[REVIEW]: PerMetrics: A Framework of Performance Metrics for Machine Learning Models
**Submitting author:** @thieu1995 (<a href="http://orcid.org/0000-0001-9994-8747">Nguyen Van Thieu</a>)
**Repository:** https://github.com/thieu1995/permetrics
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v2.0.0

@CeciliaCoelho & @KennethEnevoldsen & @ritika-giri & @GISWLH, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
