### 2023-12-13 #6143 @KennethEnevoldsen
I have gone through the package and paper, so far everything looks good, but I have a few comments:

1) The repository seems to be missing a a community guideline, including how to contribute and e.g. run the tests. I got them to run, but others might not.
2) It seems like the tests only include test for a minority of the metrics. Thus it is very hard to evaluate if the metrics are calculated correctly. I would love to see the test expanded to especially ensure agreement with existing libraries where appropriate and tests more deliberate tests of metrics which are unique to the library. Seems like these already exist in the
3) This software adds multiple improvements to the existing metrics in the sklearn library. I would have love to see at least some of these ported to the sklearn library. Is there a reason why that is not the case? (as I understand joss accept papers on contributions to existing software libraries as well)
4) I would love to see -  as one of the main example in the documentation -  an example of a metric which is missing or improved over sklearn. Including an introduction of the metric. There is a couple of such examples in the examples folder, but I would write one out in the main documentation as this is primary reason for the package.

A few minor things which is probably a preference (which I leave up to the author to decide so feel free to ignore it):
- there is  dependencies in both readme and the setup.py (I would remove the ones from the readme)
- the install instruction specifies a specific version in the readme, I would avoid that
- There is documentation is spread over the readme, documentation and in the examples/*. This might lead to some missing important elements of the docs. I would move examples and table from the readme to the documentation and only keep a minimal example in the readme.
- A lot of the readme contain links, it would make it easier to get an overview if hyperlinks were used
- there is a run.py script which seems like it could be deleted
- version are specified in multiple places, I would make sure that there is only one source truth (and the rest import from that source)
