### 2023-12-18 #6143 @GISWLH
Thanks to the editor for the invitation. This python package is excellent, offering a vast and unified of machine learning statistic metrics. I've briefly reviewed the paper and code. The main contribution of this article lies in systematically reviewing the definitions and computational differences of R, R2, and R^2 across various software libraries, providing consistent and accurate calculations. I am eager to see the paper and code published in the JOSS journal. Over the next few days, I will review the code in detail. Should any issues arise, I will raise them in the repository. Thanks for the authors contribution
