### 2024-01-08 #6143 @galessiorob
Happy New Year!

@CeciliaCoelho and @GISWLH mind updating us on your ETA for the review? Please let me know if you have any questions about the process.

@thieu1995 please review @KennethEnevoldsen [comments](https://github.com/openjournals/joss-reviews/issues/6143#issuecomment-1853712541) and address as applicable.
