### 2024-01-11 #6143 @GISWLH
Dear @galessiorob and @thieu1995 ,

Apologies for the delayed response. Over the past few weeks, I reviewed the source code. The primary recommendations have already been provided earlier; the following are some minor observations.

* Handling of NaN Values: There seems to be a lack of precise consideration for NaN values, which is crucial. When data contains empty values, some other computation packages return NaN (numpy, sklearn), aligning with our understanding. At least, a detailed explanation of how NaN is handled should be included in the README section, even though I have noticed some conditional checks in the source code.

* Returning Decimal Places: The necessity of returning decimal places, requiring users to input additional numbers, seems to add complexity. While this is just a suggestion, the use of a simple ``round`` function by the author is noted. I believe data scientists can handle these adjustments as needed.

* Coefficient of Determination and Nash Efficiency Coefficient: These should be the same, even if the source code written by the author differs. They are potentially confusing. It would be beneficial if the author could add clarifications on these in the documentation.

* Definition of Performance Metrics: When discussing various performance metrics, please ensure that each metric is clearly defined. This may involve a brief explanation or an example of where each metric is most applicable.

* Paper Organization and Impact: The paper is well-organized. I suggest reinforcing the conclusion by stating the potential impact of PerMetrics in the field of machine learning.

* References Format: The format of the references is not uniform, and some are missing DOI links.

I highly recommend that the paper be published in the JOSS journal. The author may consider the aforementioned suggestions selectively and possibly make improvements in the upcoming versions.
