### 2024-01-13 f2499b0
Fix bug in condition check
