+++ README.md
@@ -27,156 +27,11 @@ alt="PERMETRICS"/>
 
 PerMetrics is a python library for performance metrics of machine learning models. We aim to implement all 
 performance metrics for problems such as regression, classification, clustering, ... problems. Helping users in all 
-field access metrics as fast as possible
-
-* **Free software:** GNU General Public License (GPL) V3 license
-* **Total metrics**: 111 (47 regression metrics, 20 classification metrics, 44 clustering metrics)
-* **Documentation:** https://permetrics.readthedocs.io/en/latest/
-* **Python versions:** >= 3.7.x
-* **Dependencies:** numpy, scipy
+field access metrics as fast as possible. The number of available metrics include **111 (47 regression metrics, 20 classification metrics, 44 clustering 
+metrics)**
 
 
-# Notification
-
-* **Currently, there is a huge misunderstanding among frameworks around the world about the notation of R, R2, and R^2.** 
-* Please read the file [R-R2-Rsquared.docx](https://github.com/thieu1995/permetrics/blob/master/R-R2-Rsquared.docx) to understand the differences between them and why there is such confusion.
-
-<p align="center"><img src=".github/assets/rr2.png" alt="R" title="R"/></p>
-
-* **My recommendation is to denote the Coefficient of Determination as COD or R2, while the squared Pearson's 
-  Correlation Coefficient should be denoted as R^2 or RSQ (as in Excel software).**
-
-
-# Installation
-
-### Install with pip
-Install the [current PyPI release](https://pypi.python.org/pypi/permetrics):
-```sh 
-$ pip install permetrics==1.5.0
-```
-
-Or installing from the source code, use:
-```sh 
-$ git clone https://github.com/thieu1995/permetrics.git
-$ cd permetrics
-$ python setup.py install
-```
-
-Or install the development version from GitHub:
-```bash
-pip install git+https://github.com/thieu1995/permetrics
-```
-
-After installation, you can import Permetrics as any other Python module:
-
-```sh
-$ python
->>> import permetrics
->>> permetrics.__version__
-```
-
-**Let's go through some examples. The more complicated test case in the folder: examples**
-
-The [documentation](https://permetrics.readthedocs.io/) includes more detailed installation instructions and explanations.
-
-
-### Example with Regression metrics
-
-
-```python
-import numpy as np
-from permetrics import RegressionMetric
-
-## For 1-D array
-y_true = [3, -0.5, 2, 7]
-y_pred = [2.5, 0.0, 2, 8]
-
-evaluator = RegressionMetric(y_true, y_pred, decimal=5)
-print(evaluator.RMSE())
-print(evaluator.MSE())
-
-## For > 1-D array
-y_true = np.array([[0.5, 1], [-1, 1], [7, -6]])
-y_pred = np.array([[0, 2], [-1, 2], [8, -5]])
-
-evaluator = RegressionMetric(y_true, y_pred, decimal=5)
-print(evaluator.RMSE(multi_output="raw_values", decimal=5))
-print(evaluator.MAE(multi_output="raw_values", decimal=5))
-```
-
-
[…]
