+++ CONTRIBUTING.md
@@ -55,7 +55,7 @@ Here are some ways to contribute:
 
 ## Pull Requests
 
-[Pull requests](/pulls) are the best way to propose changes to the codebase. We actively welcome your pull requests:
+[Pull requests](https://github.com/thieu1995/permetrics/pulls) are the best way to propose changes to the codebase. We actively welcome your pull requests:
 
 1. Fork the repo and create your branch from `main`.
 2. If you've added code that should be tested, add tests.
@@ -65,7 +65,7 @@ Here are some ways to contribute:
 
 ## Issue Tracking
 
-We use [GitHub issues](/issues) to track public bugs and requests. Please ensure your description is clear and has 
+We use [GitHub issues](https://github.com/thieu1995/permetrics/issues) to track public bugs and requests. Please ensure your description is clear and has 
 sufficient 
 instructions to be able to reproduce the issue.
 
