### 2024-01-18 a28c4f5
Update test regression metrics
