+++ tests/test_features/test_regression.py
@@ -1,0 +1,659 @@
+#!/usr/bin/env python
+# Created by "Thieu" at 09:58, 27/07/2023 ----------%                                                                               
+#       Email: nguyenthieu2102@gmail.com            %                                                    
+#       Github: https://github.com/thieu1995        %                         
+# --------------------------------------------------%
+
+import numpy as np
+from permetrics import RegressionMetric
+import pytest
+
+
+@pytest.fixture(scope="module")  # scope: Call only 1 time at the beginning
+def data():
+    rm1 = RegressionMetric(y_true=np.array([3, -0.5, 2, 7, 5]), y_pred=np.array([2.5, 0.0, 2, 8, 5]))
+    rm2 = RegressionMetric(y_true=np.array([[3, 1], [-0.5, 2], [2, 3], [7, 4], [5, 5], [6, 6]]),
+                           y_pred=np.array([[2.5, 0.5], [0.0, 1.5], [2, 2], [8, 3], [5, 4], [3, 5]]))
+    return rm1, rm2
+
+
+def test_EVS(data):
+    rm1, rm2 = data[0], data[1]
+    # 1D
+    res = rm1.EVS()
+    assert isinstance(res, (float))
+    # ND
+    res = rm2.EVS(multi_output=None)
+    assert isinstance(res, (float))
+    res = rm2.EVS(multi_output="raw_values")
+    assert isinstance(res, (list, tuple, np.ndarray))
+    res = rm2.EVS(multi_output=(0.2, 0.7))
+    assert isinstance(res, float)
+
+
+def test_ME(data):
+    rm1, rm2 = data[0], data[1]
+    # 1D
+    res = rm1.ME()
+    assert isinstance(res, (float))
+    # ND
+    res = rm2.ME(multi_output=None)
+    assert isinstance(res, (float))
+    res = rm2.ME(multi_output="raw_values")
+    assert isinstance(res, (list, tuple, np.ndarray))
+    res = rm2.ME(multi_output=(0.2, 0.7))
+    assert isinstance(res, float)
+
+
+def test_MBE(data):
+    rm1, rm2 = data[0], data[1]
+    # 1D
+    res = rm1.MBE()
+    assert isinstance(res, (float))
+    # ND
+    res = rm2.MBE(multi_output=None)
+    assert isinstance(res, (float))
+    res = rm2.MBE(multi_output="raw_values")
+    assert isinstance(res, (list, tuple, np.ndarray))
+    res = rm2.MBE(multi_output=(0.2, 0.7))
+    assert isinstance(res, float)
+
+
+def test_MAE(data):
+    rm1, rm2 = data[0], data[1]
+    # 1D
+    res = rm1.MAE()
+    assert isinstance(res, (float))
+    # ND
+    res = rm2.MAE(multi_output=None)
+    assert isinstance(res, (float))
+    res = rm2.MAE(multi_output="raw_values")
+    assert isinstance(res, (list, tuple, np.ndarray))
+    res = rm2.MAE(multi_output=(0.2, 0.7))
+    assert isinstance(res, float)
+
+
+def test_MSE(data):
+    rm1, rm2 = data[0], data[1]
+    # 1D
+    res = rm1.MSE()
+    assert isinstance(res, (float))
+    # ND
+    res = rm2.MSE(multi_output=None)
+    assert isinstance(res, (float))
+    res = rm2.MSE(multi_output="raw_values")
+    assert isinstance(res, (list, tuple, np.ndarray))
+    res = rm2.MSE(multi_output=(0.2, 0.7))
[…]
