### 2024-02-01 8064ae0
Update format check for input data
