+++ permetrics/regression.py
@@ -14,10 +14,10 @@ class RegressionMetric(Evaluator):
     """
     Defines a RegressionMetric class that hold all regression metrics (for both regression and time-series problems)
 
-    + An extension of scikit-learn metrics section, with the addition of many more regression metrics.
-    + https://scikit-learn.org/stable/modules/model_evaluation.html#classification-metrics
-    + Some methods in scikit-learn can't generate the multi-output metrics, we re-implement all of them and allow multi-output metrics
-    + Therefore, we support calculate the multi-output metrics for all methods
+        + An extension of scikit-learn metrics section, with the addition of many more regression metrics.
+        + https://scikit-learn.org/stable/modules/model_evaluation.html#classification-metrics
+        + Some methods in scikit-learn can't generate the multi-output metrics, we re-implement all of them and allow multi-output metrics
+        + Therefore, we support calculate the multi-output metrics for all methods
 
     Parameters
     ----------
@@ -26,9 +26,6 @@ class RegressionMetric(Evaluator):
 
     y_pred: tuple, list, np.ndarray, default = None
         The prediction values.
-
-    decimal: int, default = 5
-        The number of fractional parts after the decimal point
     """
 
     SUPPORT = {
@@ -88,11 +85,10 @@ class RegressionMetric(Evaluator):
         "CRM": {"type": "min", "range": "(-inf, +inf)", "best": "0"},
     }
 
-    def __init__(self, y_true=None, y_pred=None, decimal=5, **kwargs):
-        super().__init__(y_true, y_pred, decimal, **kwargs)
+    def __init__(self, y_true=None, y_pred=None, **kwargs):
+        super().__init__(y_true, y_pred, **kwargs)
         if kwargs is None: kwargs = {}
         self.set_keyword_arguments(kwargs)
-        self.one_dim = False
 
     @staticmethod
     def get_support(name=None, verbose=True):
@@ -108,33 +104,27 @@ class RegressionMetric(Evaluator):
                 print(f"Metric {name}: {RegressionMetric.SUPPORT[name]}")
             return RegressionMetric.SUPPORT[name]
 
-    def get_processed_data(self, y_true=None, y_pred=None, decimal=None, **kwargs):
+    def get_processed_data(self, y_true=None, y_pred=None, **kwargs):
         """
         Args:
             y_true (tuple, list, np.ndarray): The ground truth values
             y_pred (tuple, list, np.ndarray): The prediction values
-            clean (bool): Remove all rows contain 0 value in y_pred (some methods have denominator is y_pred)
-            decimal (int): The number of fractional parts after the decimal point
 
         Returns:
             y_true_final: y_true used in evaluation process.
             y_pred_final: y_pred used in evaluation process
-            one_dim: is y_true has 1 dimensions or not
-            decimal: The number of fractional parts after the decimal point
+            n_out: Number of outputs
         """
-        decimal = self.decimal if decimal is None else decimal
         if (y_true is not None) and (y_pred is not None):
-            y_true, y_pred = du.format_regression_data_type(y_true, y_pred)
-            y_true, y_pred, one_dim = du.format_regression_data(y_true, y_pred)
+            y_true, y_pred, n_out = du.format_regression_data_type(y_true, y_pred)
         else:
             if (self.y_true is not None) and (self.y_pred is not None):
-                y_true, y_pred = du.format_regression_data_type(self.y_true, self.y_pred)
-                y_true, y_pred, one_dim = du.format_regression_data(y_true, y_pred)
+                y_true, y_pred, n_out = du.format_regression_data_type(self.y_true, self.y_pred)
             else:
[…]
