+++ docs/source/pages/regression/MSE.rst
@@ -19,13 +19,14 @@ MSE - Mean Squared Error
 
 	\text{MSE}(y, \hat{y}) = \frac{\sum_{i=0}^{N - 1} (y_i - \hat{y}_i)^2}{N}
 
-+ Best possible score is 0.0, smaller value is better. Range = [0, +inf)
-+ MSE: a risk metric corresponding to the expected value of the squared (quadratic) error or loss.
-
-
 Latex equation code::
 
 	\text{MSE}(y, \hat{y}) = \frac{\sum_{i=0}^{N - 1} (y_i - \hat{y}_i)^2}{N}
+
+
++ Best possible score is 0.0, smaller value is better. Range = [0, +inf)
++ MSE: a risk metric corresponding to the expected value of the squared (quadratic) error or loss.
++ `Link to equation <https://statisticsbyjim.com/regression/mean-squared-error-mse/>`_
 
 
 Example to use MSE metric:
@@ -40,12 +41,12 @@ Example to use MSE metric:
 	y_true = array([3, -0.5, 2, 7])
 	y_pred = array([2.5, 0.0, 2, 8])
 
-	evaluator = RegressionMetric(y_true, y_pred, decimal=5)
+	evaluator = RegressionMetric(y_true, y_pred)
 	print(evaluator.mean_squared_error())
 
 	## For > 1-D array
 	y_true = array([[0.5, 1], [-1, 1], [7, -6]])
 	y_pred = array([[0, 2], [-1, 2], [8, -5]])
 
-	evaluator = RegressionMetric(y_true, y_pred, decimal=5)
+	evaluator = RegressionMetric(y_true, y_pred)
 	print(evaluator.MSE(multi_output="raw_values"))
