### 2024-02-01 210c78b
Update NSLE metric
