### 2024-02-01 a08200f
Update SMAPE metric
