+++ docs/source/pages/regression/MAAPE.rst
@@ -34,7 +34,7 @@ MAAPE indicates better forecast accuracy. The MAAPE is commonly used in time ser
 prediction, and demand forecasting.
 
 + Best possible score is 0.0, smaller value is better. Range = [0, +inf)
-+ https://support.numxl.com/hc/en-us/articles/115001223463-MAAPE-Mean-Arctangent-Absolute-Percentage-Error
++ `Link to equation <https://support.numxl.com/hc/en-us/articles/115001223463-MAAPE-Mean-Arctangent-Absolute-Percentage-Error>`_
 
 
 Example to use MAAPE metric:
@@ -49,12 +49,12 @@ Example to use MAAPE metric:
 	y_true = array([3, -0.5, 2, 7])
 	y_pred = array([2.5, 0.0, 2, 8])
 
-	evaluator = RegressionMetric(y_true, y_pred, decimal=5)
+	evaluator = RegressionMetric(y_true, y_pred)
 	print(evaluator.mean_arctangent_absolute_percentage_error())
 
 	## For > 1-D array
 	y_true = array([[0.5, 1], [-1, 1], [7, -6]])
 	y_pred = array([[0, 2], [-1, 2], [8, -5]])
 
-	evaluator = RegressionMetric(y_true, y_pred, decimal=5)
+	evaluator = RegressionMetric(y_true, y_pred)
 	print(evaluator.MAAPE(multi_output="raw_values"))
