### 2024-02-01 cb79f3c
Add NSE metric
