### 2024-02-01 bff7194
Update NNSE metric
