### 2024-02-02 448c975
Update RegressionMetric class
