### 2024-02-02 afa10ea
Update docs for RegressionMetric
