+++ docs/source/pages/regression/VAF.rst
@@ -19,15 +19,15 @@ VAF - Variance Accounted For
 
 	\text{VAF}(y, f_i) = 100\% \times \frac{\sum_{i=1}^{n}(y_i - \bar{y})(f_i - \bar{f})}{\sum_{i=1}^{n}(y_i - \bar{y})^2}
 
+Latex equation code::
+
+	\text{VAF}(y, f_i) = 100\% \times \frac{\sum_{i=1}^{n}(y_i - \bar{y})(f_i - \bar{f})}{\sum_{i=1}^{n}(y_i - \bar{y})^2}
 
 + Variance Accounted For (VAF) is a metric used to evaluate the performance of a regression model. It measures the proportion of the total variance in the
 actual values that is accounted for by the variance in the predicted values.
 + Variance Accounted For between 2 signals (VAF): Best possible score is 100% (identical signal), bigger value is better. Range = (-inf, 100%]
-+ https://www.dcsc.tudelft.nl/~jwvanwingerden/lti/doc/html/vaf.html
++ `Link to equation <https://www.dcsc.tudelft.nl/~jwvanwingerden/lti/doc/html/vaf.html>`_
 
-Latex equation code::
-
-	\text{VAF}(y, f_i) = 100\% \times \frac{\sum_{i=1}^{n}(y_i - \bar{y})(f_i - \bar{f})}{\sum_{i=1}^{n}(y_i - \bar{y})^2}
 
 Example to use VAF metric:
 
@@ -41,12 +41,12 @@ Example to use VAF metric:
 	y_true = array([3, -0.5, 2, 7])
 	y_pred = array([2.5, 0.0, 2, 8])
 
-	evaluator = RegressionMetric(y_true, y_pred, decimal=5)
+	evaluator = RegressionMetric(y_true, y_pred)
 	print(evaluator.variance_accounted_for())
 
 	## For > 1-D array
 	y_true = array([[0.5, 1], [-1, 1], [7, -6]])
 	y_pred = array([[0, 2], [-1, 2], [8, -5]])
 
-	evaluator = RegressionMetric(y_true, y_pred, decimal=5)
+	evaluator = RegressionMetric(y_true, y_pred)
 	print(evaluator.VAF(multi_output="raw_values"))
