+++ permetrics/classification.py
@@ -63,8 +63,8 @@ class ClassificationMetric(Evaluator):
         "BSL": {"type": "min", "range": "[0, 1]", "best": "0"}
     }
 
-    def __init__(self, y_true=None, y_pred=None, decimal=5, **kwargs):
-        super().__init__(y_true, y_pred, decimal, **kwargs)
+    def __init__(self, y_true=None, y_pred=None, **kwargs):
+        super().__init__(y_true, y_pred, **kwargs)
         if kwargs is None: kwargs = {}
         self.set_keyword_arguments(kwargs)
         self.binary = True
@@ -85,20 +85,17 @@ class ClassificationMetric(Evaluator):
                 print(f"Metric {name}: {ClassificationMetric.SUPPORT[name]}")
             return ClassificationMetric.SUPPORT[name]
 
-    def get_processed_data(self, y_true=None, y_pred=None, decimal=None):
+    def get_processed_data(self, y_true=None, y_pred=None):
         """
         Args:
             y_true (tuple, list, np.ndarray): The ground truth values
             y_pred (tuple, list, np.ndarray): The prediction values
-            decimal (int, None): The number of fractional parts after the decimal point
 
         Returns:
             y_true_final: y_true used in evaluation process.
             y_pred_final: y_pred used in evaluation process
             one_dim: is y_true has 1 dimensions or not
-            decimal: The number of fractional parts after the decimal point
         """
-        decimal = self.decimal if decimal is None else decimal
         if (y_true is not None) and (y_pred is not None):
             y_true, y_pred, binary, representor = du.format_classification_data(y_true, y_pred)
         else:
@@ -106,22 +103,19 @@ class ClassificationMetric(Evaluator):
                 y_true, y_pred, binary, representor = du.format_classification_data(self.y_true, self.y_pred)
             else:
                 raise ValueError("y_true or y_pred is None. You need to pass y_true and y_pred to object creation or function called.")
-        return y_true, y_pred, binary, representor, decimal
+        return y_true, y_pred, binary, representor
 
-    def get_processed_data2(self, y_true=None, y_pred=None, decimal=None):
+    def get_processed_data2(self, y_true=None, y_pred=None):
         """
         Args:
             y_true (tuple, list, np.ndarray): The ground truth values
             y_pred (tuple, list, np.ndarray): The prediction scores
-            decimal (int, None): The number of fractional parts after the decimal point
 
         Returns:
             y_true_final: y_true used in evaluation process.
             y_pred_final: y_pred used in evaluation process
             one_dim: is y_true has 1 dimensions or not
-            decimal: The number of fractional parts after the decimal point
         """
-        decimal = self.decimal if decimal is None else decimal
         if (y_true is not None) and (y_pred is not None):
             y_true, y_pred, binary, representor = du.format_y_score(y_true, y_pred)
         else:
@@ -129,7 +123,7 @@ class ClassificationMetric(Evaluator):
                 y_true, y_pred, binary, representor = du.format_y_score(self.y_true, self.y_pred)
             else:
                 raise ValueError("y_true or y_pred is None. You need to pass y_true and y_pred to object creation or function called.")
-        return y_true, y_pred, binary, representor, decimal
+        return y_true, y_pred, binary, representor
 
     def confusion_matrix(self, y_true=None, y_pred=None, labels=None, normalize=None, **kwargs):
         """
@@ -146,11 +140,11 @@ class ClassificationMetric(Evaluator):
             imap (dict): a map between label and index of confusion matrix
             imap_count (dict): a map between label and number of true label in y_true
         """
-        y_true, y_pred, binary, representor, decimal = self.get_processed_data(y_true, y_pred, decimal=None)
[…]
