+++ permetrics/classification.py
@@ -185,7 +185,7 @@ class ClassificationMetric(Evaluator):
             result = dict([(label, item["precision"]) for label, item in metrics.items()])
         return result
 
-    def negative_predictive_value(self, y_true=None, y_pred=None, labels=None, average="macro", decimal=None, **kwargs):
+    def negative_predictive_value(self, y_true=None, y_pred=None, labels=None, average="macro", **kwargs):
         """
         Generate negative predictive value for multiple classification problem
         Higher is better (Best = 1), Range = [0, 1]
@@ -195,12 +195,11 @@ class ClassificationMetric(Evaluator):
             y_pred (tuple, list, np.ndarray): a list of integers or strings for y_pred classes
             labels (tuple, list, np.ndarray): List of labels to index the matrix. This may be used to reorder or select a subset of labels.
             average (str, None): {'micro', 'macro', 'weighted'} or None, default="macro"
-            decimal (int): The number of fractional parts after the decimal point
 
         Returns:
             npv (float, dict): the negative predictive value
         """
-        y_true, y_pred, binary, representor, decimal = self.get_processed_data(y_true, y_pred, decimal)
+        y_true, y_pred, binary, representor = self.get_processed_data(y_true, y_pred)
         matrix, imap, imap_count = cu.calculate_confusion_matrix(y_true, y_pred, labels, normalize=None)
         metrics = cu.calculate_single_label_metric(matrix, imap, imap_count)
 
@@ -210,16 +209,16 @@ class ClassificationMetric(Evaluator):
         if average == "micro":
             tp_global = tn_global = np.sum(np.diag(matrix))
             fp_global = fn_global = np.sum(matrix) - tp_global
-            npv = tn_global / (tn_global + fn_global)
+            result = tn_global / (tn_global + fn_global)
         elif average == "macro":
-            npv = np.mean(list_npv)
+            result = np.mean(list_npv)
         elif average == "weighted":
-            npv = np.dot(list_weights, list_npv) / np.sum(list_weights)
+            result = np.dot(list_weights, list_npv) / np.sum(list_weights)
         else:
-            npv = dict([(label, np.round(item["negative_predictive_value"], decimal)) for label, item in metrics.items()])
-        return npv if type(npv) == dict else np.round(npv, decimal)
+            result = dict([(label, item["negative_predictive_value"]) for label, item in metrics.items()])
+        return result
 
-    def specificity_score(self, y_true=None, y_pred=None, labels=None, average="macro", decimal=None, **kwargs):
+    def specificity_score(self, y_true=None, y_pred=None, labels=None, average="macro", **kwargs):
         """
         Generate specificity score for multiple classification problem
         Higher is better (Best = 1), Range = [0, 1]
@@ -229,12 +228,11 @@ class ClassificationMetric(Evaluator):
             y_pred (tuple, list, np.ndarray): a list of integers or strings for y_pred classes
             labels (tuple, list, np.ndarray): List of labels to index the matrix. This may be used to reorder or select a subset of labels.
             average (str, None): {'micro', 'macro', 'weighted'} or None, default="macro"
-            decimal (int): The number of fractional parts after the decimal point
 
         Returns:
             ss (float, dict): the specificity score
         """
-        y_true, y_pred, binary, representor, decimal = self.get_processed_data(y_true, y_pred, decimal)
+        y_true, y_pred, binary, representor = self.get_processed_data(y_true, y_pred)
         matrix, imap, imap_count = cu.calculate_confusion_matrix(y_true, y_pred, labels, normalize=None)
[…]
