### 2024-02-07 894c39b
Update condition check for y_score
