### 2024-02-07 1c06d65
Update condition check for y_pred in classification
