+++ tests/test_comparisons/test_sklearn_classification.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+# Created by "Thieu" at 16:14, 07/02/2024 ----------%                                                                               
+#       Email: nguyenthieu2102@gmail.com            %                                                    
+#       Github: https://github.com/thieu1995        %                         
+# --------------------------------------------------%
+
+
+from sklearn.metrics import log_loss
+print(log_loss(["spam", "ham", "ham", "spam"],
+         [[.1, .9], [.9, .1], [.8, .2], [.35, .65]]))
+
+
