+++ tests/test_features/test_classification.py
@@ -6,19 +6,211 @@
 
 import numpy as np
 from permetrics import ClassificationMetric
+import pytest
 
-np.random.seed(42)
+
+@pytest.fixture(scope="module")  # scope: Call only 1 time at the beginning
+def data():
+    cm1 = ClassificationMetric(y_true=np.array([0, 1, 0, 0, 1, 0]), y_pred=np.array([0, 1, 0, 0, 0, 1]))
+
+    # Example one-hot encoded y_true and y_pred
+    y_true = np.array([[0, 1, 0],  # Class 1
+                       [1, 0, 0],  # Class 0
+                       [0, 0, 1],  # Class 2
+                       [0, 1, 0],  # Class 1
+                       [0, 0, 1]])  # Class 2
+    y_pred = np.array([[0.1, 0.8, 0.1],  # Predicted probabilities for Class 1, Class 0, Class 2
+                       [0.7, 0.2, 0.1],
+                       [0.2, 0.3, 0.5],
+                       [0.3, 0.6, 0.1],
+                       [0.1, 0.2, 0.7]])
+    cm2 = ClassificationMetric(y_true=y_true, y_pred=y_pred)
+
+    y_true = np.array([0, 1, 2, 0, 2])  # Class 2
+    y_pred = np.array([[0.1, 0.8, 0.1],  # Predicted probabilities for Class 1, Class 0, Class 2
+                       [0.7, 0.2, 0.1],
+                       [0.2, 0.3, 0.5],
+                       [0.3, 0.6, 0.1],
+                       [0.1, 0.2, 0.7]])
+    cm3 = ClassificationMetric(y_true=y_true, y_pred=y_pred)
+    return cm1, cm2, cm3
 
 
-def test_ClassificationMetric_class():
-    y_true = [0, 1, 0, 0, 1, 0]
-    y_pred = [0, 1, 0, 0, 0, 1]
+def test_PS(data):
+    avg_paras = [None, "macro", "micro", "weighted"]
+    outs = (dict, float, float, float)
 
-    evaluator = ClassificationMetric(y_true, y_pred, decimal=5)
+    for idx, avg in enumerate(avg_paras):
+        for cm in data:
+            res = cm.PS(average=avg)
+            assert isinstance(res, outs[idx])
 
-    mcc1 = evaluator.matthews_correlation_coefficient()
-    mcc2 = evaluator.MCC()
-    mcc3 = evaluator.MCC(y_true, y_pred)
 
-    assert mcc1 == mcc2
-    assert mcc2 == mcc3
+def test_NPV(data):
+    avg_paras = [None, "macro", "micro", "weighted"]
+    outs = (dict, float, float, float)
+
+    for idx, avg in enumerate(avg_paras):
+        for cm in data:
+            res = cm.NPV(average=avg)
+            assert isinstance(res, outs[idx])
+
+
+def test_RS(data):
+    avg_paras = [None, "macro", "micro", "weighted"]
+    outs = (dict, float, float, float)
+
+    for idx, avg in enumerate(avg_paras):
+        for cm in data:
+            res = cm.RS(average=avg)
+            assert isinstance(res, outs[idx])
+
+
+def test_AS(data):
+    avg_paras = [None, "macro", "micro", "weighted"]
+    outs = (dict, float, float, float)
+
+    for idx, avg in enumerate(avg_paras):
+        for cm in data:
+            res = cm.AS(average=avg)
+            assert isinstance(res, outs[idx])
+
+
+def test_F1S(data):
+    avg_paras = [None, "macro", "micro", "weighted"]
+    outs = (dict, float, float, float)
+
+    for idx, avg in enumerate(avg_paras):
+        for cm in data:
[…]
