### 2024-02-20 d19ed59
Update Evaluator class
