+++ permetrics/clustering.py
@@ -31,11 +31,8 @@ class ClusteringMetric(Evaluator):
     X: tuple, list, np.ndarray, default = None
         The features of datasets. This is for calculating internal metrics
 
-    decimal: int, default = 5
-        The number of fractional parts after the decimal point
-
     raise_error: bool, default = False
-        Some metrics can't be calculate when some problems occur, show it will raise error as usual.
+        Some metrics can't be calculated when some problems occur, show it will raise error as usual.
         We can return the biggest value or smallest value depend on metric instead of raising error.
 
     biggest_value: float, default = None
@@ -94,16 +91,14 @@ class ClusteringMetric(Evaluator):
         "GPS": {"type": "min", "range": "[0, 1]", "best": "0"},
     }
 
-    def __init__(self, y_true=None, y_pred=None, X=None, decimal=5,
-                 raise_error=False, biggest_value=None, smallest_value=None, **kwargs):
-        super().__init__(y_true, y_pred, decimal, **kwargs)
+    def __init__(self, y_true=None, y_pred=None, X=None, force_finite=True, finite_value=None, **kwargs):
+        super().__init__(y_true, y_pred, **kwargs)
         if kwargs is None: kwargs = {}
         self.set_keyword_arguments(kwargs)
         self.X = X
         self.le = None
-        self.raise_error = raise_error
-        self.biggest_value = np.inf if biggest_value is None else biggest_value
-        self.smallest_value = -np.inf if smallest_value is None else smallest_value
+        self.force_finite = force_finite
+        self.finite_value = finite_value
 
     @staticmethod
     def get_support(name=None, verbose=True):
@@ -150,26 +145,29 @@ class ClusteringMetric(Evaluator):
                 y_true, y_pred, self.le = du.format_external_clustering_data(y_true, y_pred)
         return y_true, y_pred, self.le, decimal
 
-    def get_processed_internal_data(self, y_pred=None, decimal=None):
+    def get_processed_internal_data(self, y_pred=None, force_finite=None, finite_value=None):
         """
         Args:
             y_pred (tuple, list, np.ndarray): The prediction values
-            decimal (int, None): The number of fractional parts after the decimal point
+            force_finite (bool): Make result as finite number
+            finite_value (float): The value that used to replace the infinite value or NaN value.
 
         Returns:
             y_pred_final: y_pred used in evaluation process
             le: label encoder object
-            decimal: The number of fractional parts after the decimal point
+            force_finite
+            finite_value
         """
-        decimal = self.decimal if decimal is None else decimal
+        force_finite = self.force_finite if force_finite is None else force_finite
+        finite_value = self.finite_value if finite_value is None else finite_value
         if y_pred is None:              # Check for function called
-            if self.y_pred is None:     # Check for object of class called
+            if self.y_pred is None:     # Check for instance called
                 raise ValueError("You need to pass y_pred to calculate external clustering metrics.")
             else:
                 y_pred, self.le = du.format_internal_clustering_data(self.y_pred)
         else:   # This is for function called, it will override object of class called
             y_pred, self.le = du.format_internal_clustering_data(y_pred)
-        return y_pred, self.le, decimal
+        return y_pred, self.le, force_finite, finite_value
 
     def check_X(self, X):
         if X is None:
@@ -179,7 +177,7 @@ class ClusteringMetric(Evaluator):
                 return self.X
         return X
 
-    def ball_hall_index(self, X=None, y_pred=None, decimal=None, **kwargs):
[…]
