+++ permetrics/utils/cluster_util.py
@@ -132,21 +132,21 @@ def compute_confusion_matrix(y_true, y_pred, normalize=False):
     return res
 
 
-def calculate_sum_squared_error_index(X=None, y_pred=None, decimal=6):
+def calculate_sum_squared_error_index(X=None, y_pred=None):
     centers, _ = compute_barycenters(X, y_pred)
     centroid_distances = centers[y_pred]
     squared_distances = np.sum((X - centroid_distances) ** 2, axis=1)
-    return np.round(np.sum(squared_distances), decimal)
+    return np.sum(squared_distances)
 
 
-def calculate_mean_squared_error_index(X=None, y_pred=None, decimal=6):
+def calculate_mean_squared_error_index(X=None, y_pred=None):
     centers, _ = compute_barycenters(X, y_pred)
     centroid_distances = centers[y_pred]
     squared_distances = np.sum((X - centroid_distances) ** 2, axis=1)
-    return np.round(np.mean(squared_distances), decimal)
+    return np.mean(squared_distances)
 
 
-def calculate_ball_hall_index(X=None, y_pred=None, decimal=6):
+def calculate_ball_hall_index(X=None, y_pred=None):
     n_clusters = len(np.unique(y_pred))
     wgss = 0
     ## For each cluster, find the centroid and then the within-group SSE
@@ -155,17 +155,27 @@ def calculate_ball_hall_index(X=None, y_pred=None, decimal=6):
         cluster_k = X[centroid_mask]
         centroid = np.mean(cluster_k, axis=0)
         wgss += np.sum((cluster_k - centroid) ** 2)
-    return np.round(wgss / n_clusters, decimal)
+    return wgss / n_clusters
 
 
-def calculate_calinski_harabasz_index(X=None, y_pred=None, decimal=6, raise_error=True, raise_value=0.0):
+def calculate_calinski_harabasz_index(X=None, y_pred=None, force_finite=True, finite_value=0.0):
+    """
+    Args:
+        X: The X matrix features
+        y_pred: The predicted results
+        force_finite: Make result as finite number
+        finite_value: The value that used to replace the infinite value or NaN value.
+
+    Returns:
+        The Calinski Harabasz Index
+    """
     n_samples, _ = X.shape
     n_clusters = len(np.unique(y_pred))
     if n_clusters == 1:
-        if raise_error:
-            raise ValueError("The Calinski-Harabasz index is undefined when y_pred has only 1 cluster.")
+        if force_finite:
+            return finite_value
         else:
-            return raise_value
+            raise ValueError("The Calinski-Harabasz index is undefined when y_pred has only 1 cluster.")
     overall_mean = np.mean(X, axis=0)
     # Calculate between-cluster variance and cluster sizes
     cluster_sizes = np.bincount(y_pred, minlength=n_clusters)
@@ -175,50 +185,50 @@ def calculate_calinski_harabasz_index(X=None, y_pred=None, decimal=6, raise_erro
     within_var = np.sum((X - cluster_means[y_pred]) ** 2)
     # Calculate the CH Index
     res = (between_var / within_var) * ((n_samples - n_clusters) / (n_clusters - 1))
-    return np.round(res, decimal)
+    return res
 
 
-def calculate_xie_beni_index(X=None, y_pred=None, decimal=6, raise_error=True, raise_value=np.inf):
+def calculate_xie_beni_index(X=None, y_pred=None, force_finite=True, finite_value=1e10):
     n_clusters = len(np.unique(y_pred))
     if n_clusters == 1:
-        if raise_error:
-            raise ValueError("The Xie-Beni index is undefined when y_pred has only 1 cluster.")
+        if force_finite:
+            return finite_value
         else:
-            return raise_value
[…]
