+++ permetrics/clustering.py
@@ -31,17 +31,11 @@ class ClusteringMetric(Evaluator):
     X: tuple, list, np.ndarray, default = None
         The features of datasets. This is for calculating internal metrics
 
-    raise_error: bool, default = False
-        Some metrics can't be calculated when some problems occur, show it will raise error as usual.
-        We can return the biggest value or smallest value depend on metric instead of raising error.
+    force_finite: bool, default = True
+        When result is not finite, it can be NaN or Inf. Their result will be replaced by `finite_value`
 
-    biggest_value: float, default = None
-        The biggest value will be returned for metric that has min characteristic and ``raise_error=True``
-        Default = None, then the value will be ``np.inf``.
-
-    smallest_value: float, default = None
-        The smallest value will be returned for metric that has max characteristic and ``raise_error=True``
-        Default = None, then the value will be ``-np.inf``.
+    finite_value: float, default = None
+        The value that used to replace the infinite value or NaN value.
     """
 
     SUPPORT = {
@@ -114,20 +108,23 @@ class ClusteringMetric(Evaluator):
                 print(f"Metric {name}: {ClusteringMetric.SUPPORT[name]}")
             return ClusteringMetric.SUPPORT[name]
 
-    def get_processed_external_data(self, y_true=None, y_pred=None, decimal=None):
+    def get_processed_external_data(self, y_true=None, y_pred=None, force_finite=None, finite_value=None):
         """
         Args:
             y_true (tuple, list, np.ndarray): The ground truth values
             y_pred (tuple, list, np.ndarray): The prediction values
-            decimal (int, None): The number of fractional parts after the decimal point
+            force_finite (bool): Force the result as finite number
+            finite_value (float): The finite number
 
         Returns:
             y_true_final: y_true used in evaluation process.
             y_pred_final: y_pred used in evaluation process
             le: label encoder object
-            decimal: The number of fractional parts after the decimal point
+            force_finite: Force the result as finite number
+            finite_value: The finite number
         """
-        decimal = self.decimal if decimal is None else decimal
+        force_finite = self.force_finite if force_finite is None else force_finite
+        finite_value = self.finite_value if finite_value is None else finite_value
         if y_pred is None:              # Check for function called
             if self.y_pred is None:     # Check for object of class called
                 raise ValueError("You need to pass y_true and y_pred to calculate external clustering metrics.")
@@ -143,7 +140,7 @@ class ClusteringMetric(Evaluator):
                 raise ValueError("You need to pass y_true and y_pred to calculate external clustering metrics.")
             else:
                 y_true, y_pred, self.le = du.format_external_clustering_data(y_true, y_pred)
-        return y_true, y_pred, self.le, decimal
+        return y_true, y_pred, self.le, force_finite, finite_value
 
     def get_processed_internal_data(self, y_pred=None, force_finite=None, finite_value=None):
         """
@@ -544,7 +541,7 @@ class ClusteringMetric(Evaluator):
         y_pred, _, force_finite, finite_value = self.get_processed_internal_data(y_pred, force_finite, finite_value)
         return cu.calculate_hartigan_index(X, y_pred, force_finite, finite_value)
 
-    def mutual_info_score(self, y_true=None, y_pred=None, decimal=None, **kwargs):
+    def mutual_info_score(self, y_true=None, y_pred=None, **kwargs):
         """
         Computes the Mutual Information score between two clusterings.
[…]
