+++ permetrics/clustering.py
@@ -328,7 +328,6 @@ class ClusteringMetric(Evaluator):
             X (array-like of shape (n_samples, n_features)):
                 A list of `n_features`-dimensional data points. Each row corresponds to a single data point.
             y_pred (array-like of shape (n_samples,)): Predicted labels for each sample.
-            decimal (int): The number of fractional parts after the decimal point
             use_normalized (bool): We normalize the scatter matrix before calculate the Det to reduce the value, default=True
 
         Returns:
@@ -751,7 +750,7 @@ class ClusteringMetric(Evaluator):
         y_true, y_pred, _, _, _ = self.get_processed_external_data(y_true, y_pred)
         return cu.calculate_czekanowski_dice_score(y_true, y_pred)
 
-    def hubert_gamma_score(self, y_true=None, y_pred=None, decimal=None, **kwargs):
+    def hubert_gamma_score(self, y_true=None, y_pred=None, force_finite=True, finite_value=-1.0, **kwargs):
         """
         Computes the Hubert Gamma score between two clusterings.
         Bigger is better (Best = 1), Range=[-1, +1]
@@ -763,15 +762,16 @@ class ClusteringMetric(Evaluator):
         Args:
             y_true (array-like): The true labels for each sample.
             y_pred (array-like): The predicted cluster labels for each sample.
-            decimal (int): The number of fractional parts after the decimal point
+            force_finite (bool): Make result as finite number
+            finite_value (float): The value that used to replace the infinite value or NaN value.
 
         Returns:
             result (float): The Hubert Gamma score
         """
-        y_true, y_pred, _, decimal = self.get_processed_external_data(y_true, y_pred, decimal)
-        return cu.calculate_hubert_gamma_score(y_true, y_pred, decimal, self.raise_error, -1.0)
+        y_true, y_pred, _, force_finite, finite_value = self.get_processed_external_data(y_true, y_pred, force_finite, finite_value)
+        return cu.calculate_hubert_gamma_score(y_true, y_pred, force_finite, finite_value)
 
-    def jaccard_score(self, y_true=None, y_pred=None, decimal=None, **kwargs):
+    def jaccard_score(self, y_true=None, y_pred=None, **kwargs):
         """
         Computes the Jaccard score between two clusterings.
         Bigger is better (Best = 1), Range = [0, 1]
@@ -786,15 +786,14 @@ class ClusteringMetric(Evaluator):
         Args:
             y_true (array-like): The true labels for each sample.
             y_pred (array-like): The predicted cluster labels for each sample.
-            decimal (int): The number of fractional parts after the decimal point
 
         Returns:
             result (float): The Jaccard score
         """
-        y_true, y_pred, _, decimal = self.get_processed_external_data(y_true, y_pred, decimal)
-        return cu.calculate_jaccard_score(y_true, y_pred, decimal)
+        y_true, y_pred, _, _, _ = self.get_processed_external_data(y_true, y_pred)
+        return cu.calculate_jaccard_score(y_true, y_pred)
 
-    def kulczynski_score(self, y_true=None, y_pred=None, decimal=None, **kwargs):
+    def kulczynski_score(self, y_true=None, y_pred=None, **kwargs):
         """
         Computes the Kulczynski score between two clusterings.
         Bigger is better (Best = 1), Range = [0, 1]
@@ -806,15 +805,14 @@ class ClusteringMetric(Evaluator):
         Args:
             y_true (array-like): The true labels for each sample.
             y_pred (array-like): The predicted cluster labels for each sample.
-            decimal (int): The number of fractional parts after the decimal point
 
         Returns:
[…]
