+++ permetrics/utils/cluster_util.py
@@ -642,73 +642,73 @@ def calculate_czekanowski_dice_score(y_true=None, y_pred=None):
     return 2 * yy / (2 * yy + yn + ny)
 
 
-def calculate_hubert_gamma_score(y_true=None, y_pred=None, decimal=6, raise_error=True, raise_value=-1.0):
+def calculate_hubert_gamma_score(y_true=None, y_pred=None, force_finite=True, finite_value=-1.0):
     n_clusters = len(np.unique(y_pred))
     if n_clusters == 1:
-        if raise_error:
-            raise ValueError("The Hubert Gamma score is undefined when y_pred has only 1 cluster.")
+        if force_finite:
+            return finite_value
         else:
-            return raise_value
+            raise ValueError("The Hubert Gamma score is undefined when y_pred has only 1 cluster.")
     yy, yn, ny, nn = compute_confusion_matrix(y_true, y_pred, normalize=True)
     NT = yy + yn + ny + nn
     res = (NT*yy - (yy+yn)*(yy+ny)) / np.sqrt((yy+yn)*(yy+ny)*(nn+yn)*(nn+ny))
-    return np.round(res, decimal)
+    return res
 
 
-def calculate_jaccard_score(y_true=None, y_pred=None, decimal=6):
+def calculate_jaccard_score(y_true=None, y_pred=None):
     yy, yn, ny, nn = compute_confusion_matrix(y_true, y_pred, normalize=True)
-    return np.round(yy / (yy + yn + ny), decimal)
+    return yy / (yy + yn + ny)
 
 
-def calculate_kulczynski_score(y_true=None, y_pred=None, decimal=6):
+def calculate_kulczynski_score(y_true=None, y_pred=None):
     yy, yn, ny, nn = compute_confusion_matrix(y_true, y_pred, normalize=True)
     res = 0.5 * ((yy / (yy + ny)) + (yy / (yy + yn)))
-    return np.round(res, decimal)
+    return res
 
 
-def calculate_mc_nemar_score(y_true=None, y_pred=None, decimal=6):
+def calculate_mc_nemar_score(y_true=None, y_pred=None):
     yy, yn, ny, nn = compute_confusion_matrix(y_true, y_pred, normalize=True)
-    return np.round((nn - ny) / np.sqrt(nn + ny), decimal)
+    return (nn - ny) / np.sqrt(nn + ny)
 
 
-def calculate_phi_score(y_true=None, y_pred=None, decimal=6, raise_error=True, raise_value=-np.inf):
+def calculate_phi_score(y_true=None, y_pred=None, force_finite=True, finite_value=-1e10):
     n_clusters = len(np.unique(y_pred))
     if n_clusters == 1:
-        if raise_error:
-            raise ValueError("The Phi score is undefined when y_pred has only 1 cluster.")
+        if force_finite:
+            return finite_value
         else:
-            return raise_value
+            raise ValueError("The Phi score is undefined when y_pred has only 1 cluster.")
     yy, yn, ny, nn = compute_confusion_matrix(y_true, y_pred, normalize=True)
     numerator = yy * nn - yn * ny
     denominator = (yy + yn) * (yy + ny) * (yn + nn) * (ny + nn)
-    return np.round(numerator / denominator, decimal)
+    return numerator / denominator
 
 
-def calculate_rogers_tanimoto_score(y_true=None, y_pred=None, decimal=6):
+def calculate_rogers_tanimoto_score(y_true=None, y_pred=None):
     yy, yn, ny, nn = compute_confusion_matrix(y_true, y_pred, normalize=True)
     cc = (yy + nn) / (yy + nn + 2 * (yn + ny))
-    return np.round(cc, decimal)
+    return cc
 
 
[…]
