### 2024-02-22 752849d
Update clustering tests
