+++ tests/test_features/test_clustering.py
@@ -6,33 +6,288 @@
 
 import numpy as np
 from permetrics import ClusteringMetric
+import pytest
 
 np.random.seed(42)
 
 
-def test_internal_ClusteringMetric_class():
+@pytest.fixture(scope="module")
+def internal_model():
     # generate sample data
     X = np.random.uniform(-1, 10, size=(300, 6))
     y_pred = np.random.randint(0, 3, size=300)
-
-    evaluator = ClusteringMetric(y_pred=y_pred, X=X, decimal=5)
-    bhi1 = evaluator.ball_hall_index()
-    bhi2 = evaluator.BHI(X=X, y_pred=y_pred)
-    bhi3 = evaluator.BHI()
-
-    assert bhi1 == bhi2
-    assert bhi2 == bhi3
+    evaluator = ClusteringMetric(y_pred=y_pred, X=X, force_finite=True)
+    return evaluator
 
 
-def test_external_ClusteringMetric_class():
+@pytest.fixture(scope="module")
+def external_model():
     # generate sample data
     y_true = np.random.randint(0, 3, size=300)
     y_pred = np.random.randint(0, 3, size=300)
+    evaluator = ClusteringMetric(y_true=y_true, y_pred=y_pred, force_finite=True)
+    return evaluator
 
-    evaluator = ClusteringMetric(y_true=y_true, y_pred=y_pred, decimal=5)
-    prs1 = evaluator.precision_score()
-    prs2 = evaluator.PrS(y_true=y_true, y_pred=y_pred)
-    prs3 = evaluator.PrS()
 
-    assert prs1 == prs2
-    assert prs2 == prs3
+def test_BHI(internal_model):
+    evaluator = internal_model
+    res = evaluator.BHI()
+    assert isinstance(res, (float, int))
+
+
+def test_XBI(internal_model):
+    evaluator = internal_model
+    res = evaluator.XBI()
+    assert isinstance(res, (float, int))
+
+
+def test_DBI(internal_model):
+    evaluator = internal_model
+    res = evaluator.DBI()
+    assert isinstance(res, (float, int))
+
+
+def test_BRI(internal_model):
+    evaluator = internal_model
+    res = evaluator.BRI()
+    assert isinstance(res, (float, int))
+
+
+def test_KDI(internal_model):
+    evaluator = internal_model
+    res = evaluator.KDI()
+    assert isinstance(res, (float, int))
+
+
+def test_DRI(internal_model):
+    evaluator = internal_model
+    res = evaluator.DRI()
+    assert isinstance(res, (float, int))
+
+
+def test_DI(internal_model):
+    evaluator = internal_model
+    res = evaluator.DI()
+    assert isinstance(res, (float, int))
+
+
+def test_CHI(internal_model):
+    evaluator = internal_model
+    res = evaluator.CHI()
+    assert isinstance(res, (float, int))
+
+
+def test_LDRI(internal_model):
+    evaluator = internal_model
+    res = evaluator.LDRI()
+    assert isinstance(res, (float, int))
+
+
+def test_LSRI(internal_model):
+    evaluator = internal_model
+    res = evaluator.LSRI()
+    assert isinstance(res, (float, int))
+
+
+def test_SI(internal_model):
+    evaluator = internal_model
+    res = evaluator.SI()
+    assert isinstance(res, (float, int))
+
+
+def test_SSEI(internal_model):
+    evaluator = internal_model
+    res = evaluator.SSEI()
+    assert isinstance(res, (float, int))
+
+
+def test_MSEI(internal_model):
+    evaluator = internal_model
+    res = evaluator.MSEI()
+    assert isinstance(res, (float, int))
+
+
+def test_DHI(internal_model):
[…]
