+++ docs/source/pages/classification/NPV.rst
@@ -45,7 +45,7 @@ Example:
 	# y_true = ["cat", "ant", "cat", "cat", "ant", "bird", "bird", "bird"]
 	# y_pred = ["ant", "ant", "cat", "cat", "ant", "cat", "bird", "ant"]
 
-	cm = ClassificationMetric(y_true, y_pred, decimal = 5)
+	cm = ClassificationMetric(y_true, y_pred)
 
 	print(cm.npv(average=None))
 	print(cm.NPV(average="micro"))
