+++ docs/source/pages/classification/F1S.rst
@@ -50,7 +50,7 @@ Example:
 	# y_true = ["cat", "ant", "cat", "cat", "ant", "bird", "bird", "bird"]
 	# y_pred = ["ant", "ant", "cat", "cat", "ant", "cat", "bird", "ant"]
 
-	cm = ClassificationMetric(y_true, y_pred, decimal = 5)
+	cm = ClassificationMetric(y_true, y_pred)
 
 	print(cm.f1_score(average=None))
 	print(cm.F1S(average="micro"))
