+++ docs/source/pages/classification/GINI.rst
@@ -56,7 +56,7 @@ Example:
 	# y_true = ["cat", "ant", "cat", "cat", "ant", "bird", "bird", "bird"]
 	# y_pred = ["ant", "ant", "cat", "cat", "ant", "cat", "bird", "ant"]
 
-	cm = ClassificationMetric(y_true, y_pred, decimal=5)
+	cm = ClassificationMetric(y_true, y_pred)
 	print(cm.gini_index(average=None))
 	print(cm.GINI()
 	print(cm.GINI()
