+++ examples/clustering/06_speed_up_PuS.py
@@ -29,10 +29,10 @@ data, y_true, y_pred, centers = generate_dataset(num_samples, num_features, num_
 
 
 time02 = time.perf_counter()
-cm = ClusteringMetric(y_true=y_true, y_pred=y_pred, decimal=8)
+cm = ClusteringMetric(y_true=y_true, y_pred=y_pred)
 res = cm.purity_score()
 print("res: ", res, time.perf_counter() - time02 )
 
 time03 = time.perf_counter()
-s3 = cut.calculate_purity_score(y_true, y_pred, decimal=8)
+s3 = cut.calculate_purity_score(y_true, y_pred)
 print("res: ", s3, time.perf_counter() - time03)
