+++ examples/classification/03_multiple_metrics.py
@@ -13,7 +13,7 @@ from permetrics.classification import ClassificationMetric
 y_true = [0, 1, 0, 0, 1, 0]
 y_pred = [0, 1, 0, 0, 0, 1]
 
-evaluator = ClassificationMetric(y_true, y_pred, decimal=5)
+evaluator = ClassificationMetric(y_true, y_pred)
 
 ## Define list of metrics you want to use
 
