+++ examples/clustering/06_speed_up_TS-GAS-GPS.py
@@ -29,30 +29,30 @@ data, y_true, y_pred, centers = generate_dataset(num_samples, num_features, num_
 
 
 # time02 = time.perf_counter()
-# cm = ClusteringMetric(y_true=y_true, y_pred=y_pred, decimal=8)
+# cm = ClusteringMetric(y_true=y_true, y_pred=y_pred)
 # res = cm.tau_score()
 # print("res: ", res, time.perf_counter() - time02 )
 #
 # time03 = time.perf_counter()
-# s3 = cut.calculate_tau_score(y_true, y_pred, decimal=8)
+# s3 = cut.calculate_tau_score(y_true, y_pred)
 # print("res: ", s3, time.perf_counter() - time03)
 
 
 # time02 = time.perf_counter()
-# cm = ClusteringMetric(y_true=y_true, y_pred=y_pred, decimal=8)
+# cm = ClusteringMetric(y_true=y_true, y_pred=y_pred)
 # res = cm.gamma_score()
 # print("res: ", res, time.perf_counter() - time02 )
 #
 # time03 = time.perf_counter()
-# s3 = cut.calculate_gamma_score(y_true, y_pred, decimal=8)
+# s3 = cut.calculate_gamma_score(y_true, y_pred)
 # print("res: ", s3, time.perf_counter() - time03)
 
 
 time02 = time.perf_counter()
-cm = ClusteringMetric(y_true=y_true, y_pred=y_pred, decimal=8)
+cm = ClusteringMetric(y_true=y_true, y_pred=y_pred)
 res = cm.gplus_score()
 print("res: ", res, time.perf_counter() - time02 )
 
 time03 = time.perf_counter()
-s3 = cut.calculate_gplus_score(y_true, y_pred, decimal=8)
+s3 = cut.calculate_gplus_score(y_true, y_pred)
 print("res: ", s3, time.perf_counter() - time03)
