+++ requirements.txt
@@ -3,3 +3,4 @@ scipy>=1.7.1
 pytest==7.1.2
 pytest-cov==4.0.0
 flake8>=4.0.1
+scikit-learn>=1.1.2
