+++ tests/test_comparisons/test_sklearn_classification.py
@@ -5,9 +5,10 @@
 # --------------------------------------------------%
 
 import numpy as np
-from permetrics import ClassificationMetric
 import pytest
 from sklearn.metrics import accuracy_score, f1_score, fbeta_score, precision_score, recall_score
+
+from permetrics import ClassificationMetric
 
 
 @pytest.fixture(scope="module")  # scope: Call only 1 time at the beginning
@@ -18,23 +19,23 @@ def data():
 
     # Example one-hot encoded y_true and y_pred
     y_true2 = np.array([[0, 1, 0],  # Class 1
-                       [1, 0, 0],  # Class 0
-                       [0, 0, 1],  # Class 2
-                       [0, 1, 0],  # Class 1
-                       [0, 0, 1]])  # Class 2
+                        [1, 0, 0],  # Class 0
+                        [0, 0, 1],  # Class 2
+                        [0, 1, 0],  # Class 1
+                        [0, 0, 1]])  # Class 2
     y_pred2 = np.array([[0.1, 0.8, 0.1],  # Predicted probabilities for Class 1, Class 0, Class 2
-                       [0.7, 0.2, 0.1],
-                       [0.2, 0.3, 0.5],
-                       [0.3, 0.6, 0.1],
-                       [0.1, 0.2, 0.7]])
+                        [0.7, 0.2, 0.1],
+                        [0.2, 0.3, 0.5],
+                        [0.3, 0.6, 0.1],
+                        [0.1, 0.2, 0.7]])
     cm2 = ClassificationMetric(y_true=y_true2, y_pred=y_pred2)
 
     y_true3 = np.array([0, 1, 2, 0, 2])  # Class 2
     y_pred3 = np.array([[0.1, 0.8, 0.1],  # Predicted probabilities for Class 1, Class 0, Class 2
-                       [0.7, 0.2, 0.1],
-                       [0.2, 0.3, 0.5],
-                       [0.3, 0.6, 0.1],
-                       [0.1, 0.2, 0.7]])
+                        [0.7, 0.2, 0.1],
+                        [0.2, 0.3, 0.5],
+                        [0.3, 0.6, 0.1],
+                        [0.1, 0.2, 0.7]])
     cm3 = ClassificationMetric(y_true=y_true3, y_pred=y_pred3)
     return (y_true1, y_pred1), (y_true2, y_pred2), (y_true3, y_pred3), cm1, cm2, cm3
 
@@ -52,7 +53,6 @@ def test_AS(data):
     # res31 = cm3.PS(average="micro")
     # res32 = accuracy_score(y_true3, y_pred3)      # ValueError: Classification metrics can't handle a mix of multiclass and continuous-multioutput targets
     # assert res31 == res32
-
 
     # avg_paras = [None, "macro", "micro", "weighted"]
     # outs = (dict, float, float, float)
@@ -105,4 +105,3 @@ def test_RS(data):
     res11 = cm1.RS(average="macro")
     res12 = recall_score(y_true1, y_pred1, average="macro")
     assert res11 == res12
-
