### 2024-02-24 a2ad15e
Update build paper workflow
