+++ .github/workflows/draft-pdf.yml
@@ -17,7 +17,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: paper.md
+          paper-path: paper/source/paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -25,7 +25,7 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper.pdf
+          path: paper/paper.pdf
       - name: save pdf to repo
         uses: stefanzweifel/git-auto-commit-action@v4
         with:
