+++ paper/paper.bib
@@ -175,7 +175,8 @@
 	journal={Journal of Machine Learning Research},
 	volume={12},
 	pages={2825--2830},
-	year={2011}
+	year={2011},
+	doi={10.48550/arXiv.1201.0490}
 }
 
 @misc{benhamner,
@@ -232,13 +233,13 @@
   doi={10.48550/arXiv.1512.01274},
 }
 
-@article{desgraupes2013clustering,
-  title={Clustering indices},
-  author={Desgraupes, Bernard},
-  journal={University of Paris Ouest-Lab Modal’X},
-  volume={1},
-  number={1},
-  pages={34},
-  year={2013},
-  publisher={Paris, France:}
+@article{van2023metacluster,
+  title={MetaCluster: An open-source Python library for metaheuristic-based clustering problems},
+  author={Van Thieu, Nguyen and Oliva, Diego and P{\'e}rez-Cisneros, Marco},
+  journal={SoftwareX},
+  volume={24},
+  pages={101597},
+  year={2023},
+  publisher={Elsevier},
+  doi={10.1016/j.softx.2023.101597}
 }
