### 2024-02-24 c11a24d
Update new version for JOSS journal
