+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2021, Thieu'
 author = 'Thieu'
 
 # The full version, including alpha/beta/rc tags
-release = '1.4.3'
+release = '2.0.0'
 
 
 # -- General configuration ---------------------------------------------------
