### 2024-02-24 #6143 @thieu1995
Hi everyone, @KennethEnevoldsen , @GISWLH, @CeciliaCoelho, and @galessiorob

I have made modifications to the source code as well as the documentation based on your suggestions. The latest updated version is 2.0.0. I hope this version meets the basic requirements that everyone has put forward.

Thank you for your valuable feedbacks. I apologize for the time it took to update to the new version. The comment by @GISWLH made me reconsider the pre-check condition and parameters for performance metrics. As a result, I had to update most of the metrics in this library.

If there are any areas where you are still not satisfied, please feel free to create issue comments directly in my library.
