### 2024-02-26 #6143 @GISWLH
Upon reviewing the revisions @thieu1995 implemented based on my recommendations, I am pleased to endorse this manuscript for publication. Thank you for your contribution to the field of machine learning evaluation metrics. I believe the previous modifications were necessary to make ``PerMetrics`` be a professional and standardized evaluation metric library. I have started using the ``PerMetrics`` and I believe more and more scholar will come to use this standard library.
