### 2024-03-06 #6143 @galessiorob
🎉  @thieu1995 thanks so much for all the hard work to get this paper in publishing shape, and to the reviewers @KennethEnevoldsen , @CeciliaCoelho , @GISWLH , thank you for lending your time and expertise!

[I have put together some final suggestions for the actual paper ](https://github.com/thieu1995/permetrics/compare/master...galessiorob:permetrics:patch-1)- feel free to take as much as you see fit. After those final edits we can do a final review and recommend the paper for publication. Almost there!
