### 2024-03-07 #6143 @galessiorob
Thank you, @CeciliaCoelho, @KennethEnevoldsen, @GISWLH, for your work to get this paper ready to be published!

I have removed @ritika-giri due to lack of response (in this issue and via email)

@thieu1995 please deposit your software in Zenodo and post the DOI here and the most recent version. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing.
