### 2024-03-09 9da4424
Update CITATION file
